# internal/server/openapi.yaml
openapi: "3.1.0"
info:
  title: Dumpster API
  description: Multimodal knowledge base REST API
  version: "1.0.0"
servers:
  - url: /
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: >-
        An anonymous session cookie minted server-side on the first request.
        No login or registration is required: the API automatically issues a
        new session_id cookie whenever the request arrives without one or
        presents an expired session. The cookie drives RLS isolation so each
        session only sees its own knowledge bases, documents, and chunks.
  schemas:
...
    CommunityResult:
      type: object
      description: >-
        A knowledge base's most recent Louvain community-detection run.
        computed_at is null when detection has never been run for this
        knowledge base — a normal first-run state, not an error.
      properties:
        computed_at:
          type: string
          format: date-time
          nullable: true
        modularity:
          type: number
        community_count:
          type: integer
        node_count:
          type: integer
        edge_count:
          type: integer
      required: [computed_at, modularity, community_count, node_count, edge_count]
...
    GraphNode:
      type: object
      description: >-
        One canonical entity in a knowledge base's graph view. community_id
        is null if community detection has never been run, or ran before
        this entity existed.
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        type:
          type: string
        community_id:
          type: integer
          nullable: true
        degree:
          type: integer
        document_count:
          type: integer
          description: >-
            Number of distinct documents contributing at least one mention of
            this entity — a signal of cross-document overlap in a knowledge
            base with more than one source.
      required: [id, label, type, community_id, degree, document_count]
    GraphEdge:
      type: object
      description: One aggregated co-occurrence edge between two canonical entities.
      properties:
        source:
          type: string
          format: uuid
        target:
          type: string
          format: uuid
        weight:
          type: number
        document_count:
          type: integer
          description: >-
            Number of distinct documents whose chunks contributed a
            co-occurrence to this pair.
      required: [source, target, weight, document_count]
    GraphView:
      type: object
      description: >-
        A knowledge base's canonical-entity graph, shaped for display —
        every canonical entity as a node and every aggregated co-occurrence
        edge. Unfiltered; percentile-based server-side filtering is a later
        refinement, not yet applied here.
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
      required: [nodes, edges]
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
paths:
...
  /kbs/{id}/communities:
    get:
      summary: Get a knowledge base's most recent community-detection result
      description: >-
        Returns a zero-value result (null computed_at) rather than a 404
        when community detection has never been run for this knowledge
        base yet — that's a normal first-run state, not an error.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The knowledge base's community-detection summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Knowledge base not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Recompute community detection over a knowledge base's entity graph
      description: >-
        Manually triggered, synchronous: runs Louvain community detection
        over the canonical-entity graph and fully replaces the previous
        result. Not run automatically on document upload/delete.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The new community-detection summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Knowledge base not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: >-
            The knowledge base has too many canonical entities to compute
            communities synchronously.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "504":
          description: Community detection did not finish within the configured deadline.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /kbs/{id}/graph:
    get:
      summary: Get a knowledge base's canonical-entity graph, shaped for display
      description: >-
        Returns every canonical entity as a node (with its current
        community_id and degree) and every aggregated co-occurrence edge.
        Does not compute anything -- community_id reflects whatever the
        most recent POST /kbs/{id}/communities run assigned, null if that
        has never run.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The knowledge base's graph.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphView"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Knowledge base not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...